+++ src/App.tsx
@@ -5,7 +5,7 @@ import Footer from './components/layout/Footer'
 import IDFindPage from './pages/IDFindPage'
 import Signup from './pages/Signup'
 import Login from './pages/Login'
-import MyPage from './pages/MyPage'
+import MyPage from './pages/MyPage/MyPage'
 import MainPage from './pages/MainPage'
 import FundingPage from './pages/FundingPage'
 import Search from './pages/Search'
